Endpoint de Server

+++ src/Logger.js
@@ -3,7 +3,9 @@ const fs = require('fs');
 class Logger {
     constructor(path) {
         this.path = path;
-        fs.truncate(path, 0);
+        if(fs.existsSync(path)) {
+            fs.truncate(path, 0);
+        }
     }
 
     log(log, tag) {
